+++ Changelog.md
@@ -14,6 +14,7 @@
 #### Bugfixes
 
 - Remove range.detach() [#73](https://github.com/upfrontIO/editable.js/pull/73)
+- Separate instance and global configuration. Fixes [#75](https://github.com/upfrontIO/editable.js/issues/75)
 
 
 # v0.3.2 (2014-06-11)